pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "ale-lite"
version = "0.1.0"
description = "ALE-style lightweight agentic training ecosystem"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "ALE-lite"}]
keywords = ["agents", "rlhf", "evaluation"]
dependencies = [
  "openai>=1.50.0",
  "typer>=0.12.0",
  "pyyaml>=6.0.1",
  "pydantic>=2.7.0",
  "rich>=13.7.0",
]

[project.optional-dependencies]
train = [
  "transformers>=4.44.0",
  "trl>=0.11.0",
  "peft>=0.12.0",
  "datasets>=2.20.0",
]

dev = [
  "pytest>=8.0.0",
  "mypy>=1.10.0",
  "ruff>=0.5.5",
]

[project.scripts]
iflow = "ale_lite.iflow.cli:app"
rock = "ale_lite.rock.cli:app"
roll = "ale_lite.roll.cli:app"
tbp = "ale_lite.tbp.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/ale_lite"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true

[tool.pytest.ini_options]
addopts = "-q"
